+++ src/extensions/logic/exceptions/exceptions.py
@@ -1,0 +1,6 @@
+class InvalidInputException(Exception):
+    pass
+
+
+class InvalidFlagException(Exception):
+    pass
