+++ src/extensions/logic/lib/command.py
@@ -1,0 +1,2 @@
+class Command:
+    pass
