+++ src/extensions/logic/poll/default_values.py
@@ -1,0 +1,5 @@
+DEFAULT_DURATION = "1m"
+
+# Seconds for a period. Each period will update the discord message containing the poll and point show
+# how much time left
+REFRESH_RATE = 5
